rename camelcase files
